requirements.txt
# Requirements for IRS Tax Form Parser
# ====================================

# Core data processing
pandas==2.2.3

# Natural Language Processing  
spacy==3.8.0

# OCR functionality
pytesseract==0.3.13

# Database connectivity
mysql-connector-python==9.1.0

# Testing framework
pytest==8.3.3

# Additional dependencies for compatibility
Pillow>=8.0.0
numpy>=1.24.0

# Additional dependencies for PDF processing
pdf2image==1.17.0
Pillow==10.4.0

# Flask for API
Flask==3.0.3

# Installation Instructions:
# =========================
# 1. Install Tesseract OCR:
#    - Ubuntu/Debian: sudo apt-get install tesseract-ocr
#    - macOS: brew install tesseract
#    - Windows: Download from https://github.com/UB-Mannheim/tesseract/wiki
#
# 2. Install spaCy English model:
#    python -m spacy download en_core_web_sm
#
# 3. Install Python dependencies:
#    pip install -r requirements.txt
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and Monitoring
# ----------------------
structlog==23.2.0
loguru==0.7.2

# XML Processing (for e-filing)
# -----------------------------
lxml==4.9.3
xmltodict==0.13.0

# Date and Time Utilities
# -----------------------
python-dateutil==2.8.2
pytz==2023.3

# File Processing
# ---------------
python-magic==0.4.27
chardet==5.2.0

# Validation and Parsing
# ----------------------
marshmallow==3.20.1
jsonschema==4.20.0

# Development and Testing (optional for production)
# ------------------------------------------------
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
# -------------
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Performance and Optimization
# ----------------------------
redis==5.0.1
celery==5.3.4

# Cloud Storage (optional)
# ------------------------
boto3==1.34.0
azure-storage-blob==12.19.0
google-cloud-storage==2.10.0

# Jupyter Notebook Support (development)
# --------------------------------------
jupyter==1.0.0
ipykernel==6.26.0
matplotlib==3.8.2
seaborn==0.13.0

# API Documentation
# -----------------
swagger-ui-bundle==0.0.9

# Rate Limiting and Caching
# -------------------------
slowapi==0.1.9
cachetools==5.3.2

# File Format Support
# -------------------
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# Async Support
# -------------
asyncio==3.4.3
aiofiles==23.2.1

# Production WSGI Server
# ----------------------
gunicorn==21.2.0

# Environment specific packages
# -----------------------------
# Uncomment based on your deployment needs:

# For AWS deployment:
# boto3==1.34.0
# awscli==1.32.0

# For Azure deployment:
# azure-identity==1.15.0
# azure-keyvault-secrets==4.7.0

# For Google Cloud deployment:
# google-auth==2.25.2
# google-cloud-secret-manager==2.18.0

# For Kubernetes deployment:
# kubernetes==28.1.0

# Memory profiling (development)
# ------------------------------
memory-profiler==0.61.0
psutil==5.9.6